Remediate audit findings across ndarray, openblas, pocketfft, and resource_scope
- Fix native memory safety, 64-bit integer precision, and view lifecycle issues across C++ hooks (custom_ufuncs.cpp, custom_sorting.cpp, custom_indexing.cpp, npz_io.cpp) and FFI layers
- Add missing DType coverage (float16, bfloat16, int8, uint16, uint32, uint64, complex64, complex128) across padding, sorting, indexing, binning, bitwise, DSP, calculus, and linear algebra decompositions
- Harden out: aliasing and non-contiguous strided buffer handling across elementwise ufuncs, reductions, sorting, splitting, random generation, and linear algebra operations
- Add matrix condition number (cond), stateful RandomGenerator, and .npy v2.0 (32-bit header length) support
- Improve cross-platform build hooks for macOS/iOS Accelerate framework and out-of-tree KissFFT staging
- Add regression test suites covering all 22 review cycles

+++ pkgs/ndarray/test/math/linalg_deep_coverage_90_test.dart
@@ -902,7 +902,9 @@ void main() {
 
           final nNuc = norm(m, ord: 'nuc');
           final svdRes = svd(m);
-          final expectedNuc = svdRes.s.toList().reduce((a, b) => a + b);
+          final expectedNuc = svdRes.s.toList().reduce(
+            (a, b) => Float64(a + b),
+          );
           expect(nNuc.scalar, closeTo(expectedNuc, 1e-5));
           svdRes.dispose();
